Lumyk-copy/requirements.txt
alembic==1.15.2
aniso8601==10.0.0
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.9.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
fastapi==0.115.12
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
gunicorn
